Improve code coverage for rules (#515)
* :see_no_evil: Ignore coverage report

* :handshake: Implement common sense coverage exclusions
1. Don't need to test `sqlfluff/__main__ since it's tested by subprocess.check_output rather than imported directly

2. Ignore version checks for now, since we're not implementing any different behaviors by version. I could also exclude these lines in particular with a `pragma: no cover` if you prefer.

* :recycle: Simplify trailing whitespace fixes

This was interesting -- the `else:` was never being hit in the tests, which didn't originally make sense to me :thinking: because that should break the `while True:` condition. It must be getting broken by something else? Either way, this makes it more readable and ensures coverage

* :skull: Remove dead code from Rule_L002._eval

I tried to create test cases to hit this code, but couldn't figure it out. It seems like checking at the beginning and checking at newlines covers all use cases for this rule.

* :muscle: Improve coverage for Rule_L003._make_indent

There didn't seem to be any test cases for indent_unit = tabs, so +1 for
that. Also got rid of the error handling, which is now taken care of at
the config level

* :skull: Remove dead code from Rule_L003._coerce_indent_to

:thinking: Seems like the intention of this was to remove any hanging
indents, when an incorrect indent is found on a segement. Try as I may
I couldn't hit this line, so I'm guessing we don't need it
:man_shrugging:

* :skull: Remove dead code from Rule_L003._eval

1. `else: pass` isn't necessary
2. No need to check if a comment wasn't found, because previous if logic
   already ensures that a comment was found

* :white_check_mark: Add L008 test case comma w/o whitespace

* :bug: Allow capitalised keywords

This was actually a bug that we squashed early, since I discovered that
it wasn't being hit by tests, and then my test case failed.

Hooray for coverage!  :tada:

Also get rid of config checking, which is handled separately

* :fire: Remove uneeded check from Rule_L011

If there wasn't whitespace or a new line before an alias, then it
wouldn't register as an alias in SQL.

Unless you can think of a test case, then I think this is safe to remove

* :white_check_mark: Test allow_scalar cases for Rule_L013

* :skull: Remove dead code from Rule_L016._eval_line_for_break.Section

There was some other untested code in here, but I thought it better to
leave it in, since it's specifically meant to catch cases we can't
imagine

* :skull: Remove dead code from Rule_L017._eval

* :skull: Remove dead code from Rule_L003._eval_line_for_breaks

I made sure that getting a double comma was impossible, don't think we
need to keep this check in here

* :skull: Remove dead code from Rule_L018._eval

* :white_check_mark: Add test cases for Rule_L018

* :skull: Remove dead code from Rule_L019._last_code_seg

There is no need for this check, since we'll always check that
len(raw_stack) > 1 in self._eval

* :white_check_mark: Add test for Rule_L020.__lint_references_and_aliases

* :white_check_mark: Add test cases for Rule_L021._eval

* :white_check_mark: Add test cases for Rule_L022._eval

* :name_badge: Rename var for easier debugging

In pdb, `r` shows the current stack, and so it made it hard to inspect
this value in debugging

* :bug: Fix wildcard reference name

Score another for discovering unlinted lines :tada:

* :leftwards_arrow_with_hook: Revert f6d97ed, add test case instead"

* :man_facepalming: Add back in removed test