Config option to load libraries for jinja templating
If the user of sqlfluff is using the jinja templater for their
dbt project, there is a good chance they have dbt_utils and/or
some other libraries within their sql code. While switching to the
dbt templater may work for some, it may not work for others because
many of these libraries' macros may not be compliant with the lint
rules. Allowing the user to "mock" out these libraires makes the
sqlfluff linter usable without having to hack together some solution
that requires changing their code.