Macros Defined in Config take Precedence over Path Macros
If I define a path in my config where all my macros live,
and then define a few specific macros in the config file, the config
file macros will take precedent over the one defined in the path.
This allows the user to generally use most of their macros that
are being used in prod, and then just mock out the non compliant ones.