:bug: Account for inline comments when fixing commas
If you see the test, you'll notice there is a trailing whitespace before
the comma. This does not happen however when I run `sqlfluff fix`, I'm
guessing because the other rules remove it for me. Ideally I would like
to not have this, but I was having a hard time figuring out how to get
the comma to be place immediately _after_ the last code segment.
Thoughts are welcome!