Add handling for redshift convert function data type argument (#3653)
Fixes issue #3651.

Previously, the first argument to the `convert` redshift function was
interpreted as a column reference rather than a data type. This could
cause rule L027 to fail since the data type wasn't qualified with a
table name even though that doesn't make sense. This adds explicit
handling for `convert` to eliminate that issue.