Allow lambda argument columns to be unqualified
Fixes https://github.com/sqlfluff/sqlfluff/issues/3961

This adds code to identify column references that are part of a lambda
expression, and mark them as "standalone aliases" that are exempt from rule
L027. (I.e. they are allowed to be unqualified, because there is no further
way to qualify them.)