Fix type error in get_rules hook of plugin example
The `get_rules` function in the plugin example template had a type
error. The function returns all the rule classes that are defined in
the plugin. The return type was incorrectly annotated as
`List[BaseRule]` which means a list of *instances* of `BaseRule` and not
classes derived from `BaseRule`. Making it `List[Type[BaseRule]]`
instead fixes the error.

For reference, the type error shown by pyright was:
```
Expression of type "list[Type[Rule_Example_L001]]" cannot be assigned to return type "List[BaseRule]"
...
  "Type[type]" is incompatible with "Type[BaseRule]"
```