:bug: Check verbosity level of pytest run before running certain tests (#4167)
This test failed when I just ran `pytest`, but succeeded when I ran `pytest -vv`. This is because "select" only appeared in the output when the verbosity level was at least two. Felt like the simplest change to just double check the verbosity level of the test run before deciding to check for this regex match