Fix postgres column constraint default syntax
This updates the default part of postgres's column constraint segment
to match ansi's. Previously, this including an `ExpressionSegment`
which caused putting a `NOT NULL` after a default value to have the
not null part included as part of the default expression. This fixes
that so you can put `NOT NULL` either before or after the default
value and it will parse similarly.