Capitalization rules ignore templated code only if configured to
The pull request at https://github.com/sqlfluff/sqlfluff/pull/2566
made sense for users who set ignore_templated_areas to true because
those macros are not under their control.  However, if a user is writing
their own macros and would like to lint those, that commit made the
capitalization rules completely unusable.

For example, the following code in a single file would not trigger any
violations, even if ignore_templated_areas is set to false:

    {% macro abc() %}
        sElEcT a, B fRoM abc;
    {% endmacro %}
    {{ abc() }}

That is... a surprising result...