Fix LT01 spacing check in templated areas
When both the LT01 rule is on, and ignore_templated_areas is False,
the LT01 rule yields a large number of false positives.  For example,
a template like this:

    {{ 'SELECT 1;' }}

would fail with a false positive error like this, if you run with
ignore_templated_areas set to false:

    L:   1 | P:   1 | LT01 | Expected single whitespace between 'SELECT'
        keyword and numeric literal. [layout.spacing]

In respace.py, the original "if" block that bypasses template segments
was added recently in:

https://github.com/sqlfluff/sqlfluff/commit/0b1f2aab2ac5eceeae5c38b049d3b8e40c338ab7

That commit was mainly focused on indentation.  I'm not sure why the
change in rescape.py was made... but deleting it doesn't cause any test
cases to fail, and it helps my new test cases pass.