Make the test__parser__lexer_slicing_calls test stricter
The test__parser__lexer_slicing_calls wasn't really asserting on the
output segments, so who knows what the lexer was doing?  This commit
adds a list of expected segments that we assert against.

It is similar in nature to test__parser__lexer_obj, but this test
supports Jinja templating, and asserting on segment type and information
from the TemplateSegment.