Preserve zero-length template segments
I am making a custom templater that will intentionally yield zero length
slices with a special slice_type.  This is going to be used to signal
to some rules information about special markers in the templated result.
Unfortunately, SQLFluff strips these right out unless they are prefixed
with "block".  Thus, I had to implement this fragile and hacky
workaround from my templater when emitting slices:

[
    TemplatedFileSlice("block_start", ...),
    TemplatedFileSlice("block_my_custom_slice_type", ...),
    TemplatedFileSlice("block_end", ...),
]

Prefixing the slice type with "block" was necessary to trick this line
of code in lexer.py in _handle_zero_length_slice:

    if tfs.slice_type.startswith("block"):
        <snip>

The next problem was that the yield statement was failing:

    yield TemplateSegment.from_slice(...block_uuid=block_stack.top())

Since the block_stack was empty, this would crash.  Thus the final hack:
first emit some fake block start/end slices to put something on the
block stack, and then the custom "block" slice type in the middle.

This all seems very hacky and fragile, so I'd rather just pass through
these segments as-is without going down the "block" branch of code.

This commit removes the "if" condition that excluded these segments.
A new test function is also introduced that allows us to continue to
add more test cases where we want to specify an exact TemplatedFile
input to the lexer, rather than depending on a particular templater
implementation in the test.