Fix multiple parse issues in Expression_A_Grammar
This changes the grammar to look more like the original Cockroach Labs
diagram from the link.  Some of the test cases that the original code
struggled with:

The code would fail with a parse error on these test cases:

SELECT 1 * -(5);
SELECT 1 * - - - (5);
SELECT 1 * + + (5);
SELECT 1 * - - - func(5);
SELECT 1 * +(5);
SELECT 'abc' LIKE - - 5;

All of these except the last one are 100% valid SQL and run without
error on PG 15.  The last one parses on PG 15: it doesn't die with a
syntax error, but rather goes on to give an error about a type mismatch,
I think meaning that it was able to parse the statement.

Next, the code would "succeed" at parsing this test case:

SELECT FALSE AND NOT NOT NOT (TRUE);

Unfortunately, the middle "NOT" was parsed as a column reference.

Finally, the original motivation for this commit: the expression
"FALSE NOT NULL" will parse as a valid expression.  Well, clearly it is
not!  This might ordinarily be relatively harmless, except for the fact
that it interferes with table constraints.

In particular, I want to support expressions for column DEFAULTs.  That
is, the following should parse correctly:

    CREATE TABLE (col int DEFAULT FALSE NOT NULL)

Currently, it works because CREATE TABLE doesn't support expressions at
all.  However, if we add the ExpressionSegment to parse expression
defaults, the parser will incorrectly consider the "NOT NULL" as being
part of the DEFAULT constraint, rather than creating a separate NOT
NULL constraint segment.

In order for this change to pass tests, I had to increase the recursion
limit, since this commit restores the recursive expression grammar.  I
found it hard for me to reason about it and compare vs the Cockroach
Labs diagrams since they weren't the same.  I saw there is some concern
in the past about that, but I think it is inconsequential in reality:
the amount of memory used by the act of recursion for stack frames
appears to be a fraction of the amount of memory otherwise being
allocated getting to that binary operator.

Some of the things I tried out with ansi/expression_recursion_2.sql:

1.  To start out with, I ran both new & old versions with
    "sqlfluff parse -vvvvv", and then compared the number of lines.
    They were roughly similar, which suggests complexity has not
    changed.

2.  I added these lines to the top of the core/parser/grammar/base.py in
    the Ref.match function, and then ran both old & new versions:

    print("match frame", len(inspect.getouterframes(inspect.currentframe())))
    print("match usage", resource.getrusage(resource.RUSAGE_SELF).ru_maxrss)

    This prints both the stack depth and the memory usage on every call
    to match.  For the expression_recursion_2.sql file, the former was
    a maximum of around 1560 frame depth for the test case.  The test
    case has around 67 binary operators, which works out to around 22
    stack frames per operator, which seems... workable in practice for
    now, even if it's ugly in theory?

    For the RSS memory usage, it was an increase of about 9 MB.  This
    was a very very similar number to the baseline nonrecursive version.