Add ExpressionSegment to CREATE TABLE ... DEFAULT
When creating a table with a DEFAULT in PostgreSQL, you can specify an
expression for the default value.  It doesn't even have to be in
parenthesis.

However, the existing grammar doesn't allow for these cases.  This
commit adds ExpressionSegment to the end of the existing grammar, which
is similar to what exists for ALTER TABLE.