Keep out zero length keywords
Keep out the zero length keyword.  Their inclusion is due to the fact
that some keywords are specified in multiline strings, and the trailing
newline results in an empty string.

I think their inclusion is almost certainly accidental, so this code
strips the input string before we split it in order to avoid that
problem.