SignedSegmentGrammar: only match one sign / preserve QualifiedNumericLiteralSegment in expressions
Numeric literals only have up to one sign, not an unlimited number:

 - http://web.cecs.pdx.edu/~len/sql1999.pdf -- defines:
     <signed numeric literal> ::= [ <sign> ] <unsigned numeric literal>
     <sign> ::= <plus sign> | <minus sign>

 - PostgreSQL has a similar rule:
     https://github.com/postgres/postgres/blob/4380c2509d51febad34e1fac0cfaeb98aaa716c5/src/backend/parser/gram.y#L16782-L16788

So it's a little unusual for us to lump several signs together in the
same "numeric_literal", which at that point requires doing parsing of
unary operators just to interpret the numeric_literal (e.g. a rule could
not do "int(num_lit.raw)" to get the integer value).

This commit changes the SignedSegmentGrammar so that it only matches one
sign, not many signs.  This means the QualifiedNumericLiteralSegment
will match with one sign only.

This was required for the second part of this commit: the expression
parser will now no longer match on a unary + or - operator if it can
also be interpreted as a QualifiedNumericLiteralSegment.  This fixes a
regression from the earlier commit in this branch where we no longer
saw QualifiedNumericLiteralSegment in the parse trees.  While that was
still a valid parse tree, it's probably the case most people would
expect us to use QualifiedNumericLiteralSegment and group a single sign
with it, so we should do that.

The end result is that "SELECT 3 + - - +4;" now parses as:

    numeric_literal:                      '3'
    whitespace:                           ' '
    binary_operator:                      '+'
    whitespace:                           ' '
    sign_indicator:                       '-'
    whitespace:                           ' '
    sign_indicator:                       '-'
    whitespace:                           ' '
    numeric_literal:
        sign_indicator:                   '+'
        numeric_literal:                  '4'