Remove most expression recursion by using tail recursion
The earlier commit in this branch fixed a lot of issues, but it had the
downside of re-introducing recursion, which SQLFluff is particularly
sensitive to since it makes recursive function calls.

This commit refactors the a_expr gramamr into smaller pieces.  In
particular, the recursive element necessary to support unary operators
has been moved into its own grammar.

The idea now is that whenever a part of the grammar _ends_ with a
single recursive call to a_expr, we instead call a special
"tail recursive" grammar that includes the non-repeating elements that
prefix the a_expr (e.g. unary operators).

If the recursive call is in the middle of the grammar, such as with

    a IN (b)

then you can't use this, because you can't have more things that match
after the reference to the tail recursion grammar.  (In this case, the
closing parenthesis prevents it.)

This commit is sufficient to avoid recursion errors on the test cases
that have long expressions, so it reverts the all to
sys.setrecursionlimit.