Fix the grammar for PG CREATE EXTENSION
It can be either a non-reserved naked identifier, or a string literal.
There's no such thing as a naked 1.2.3.4 version specification in the
grammar.  If you want a specific version like that, you have to use a
string literal.

The examples in the .sql test file won't even parse on my PG instance.