Recognize "on" value and integers for PG SET statement
Normally you can't specify reserved keywords in the value for a "SET"
statement.  However, "ON" is carved out as a special exception in the
grammar.  We'll treat it as a naked_identifier so that it will be
capitalized similarly to "off", or any other non-reserved keywords or
identifiers used in the variable value.

The ALTER ROLE statement had an issue where it wouldn't accept integers.
It did accept "ON" and "OFF" but as keywords.  Also it wasn't accepting
a list when it should have.