Add support for the PG VACUUM statement
This commit adds support for the PG VACUUM statement.

Redshift already had a VACUUM statement, but it's totally different from
vanilla PG.

This one's a bit awkward because the actual upstream grammar file just
specifies a generic map of a few reserved keywords + all non-reserved
keywords + IDENT, as being valid configuration keys.  We could use the
properties_naked_identifier for this, but somehow using a keyword seems
more appropriate, as we generally would want these to be upper-case, and
most of the words are already keywords of some kind.  It does have the
downside in that PG will parse a query with a non-sensical parameter
name, whereas we won't.