Recognize quoted data types
The ANSI standard allows for quoted data types, as do many
implementations.  However, SQLFluff does not.  Therefore the following
trivial SQL would fail to parse:

    CREATE TABLE foo (
        col "my type"
    );

This commit fixes the issue by changing DatatypeIdentifierSegment so
that it can also match with SingleIdentifierGrammar.  The initial
data_type_identifier type segment is left as the default, since there
is a capitalization rule that refers to it.  However, such a rule would
be inappropriate to apply to quoted data types... so this commit just
leaves the normal quoted_identifier segment type if the data type was
quoted.

In order for this change to work, we have to move NOT from the list of
ANSI non-reserved keywords to the list of ANSI reserved keywords.
According to http://web.cecs.pdx.edu/~len/sql1999.pdf it is in fact a
reserved word.  If we don't do this, one of the test cases will fail to
parse correctly.

We also remove the TdCastIdentifierGrammar from SingleIdentifierGrammar.
It contains an entire ExpressionSegment!  That's not the right place in
the parsing to be evaluating expressions.  A glance at DatatypeSegment
suggests that it would still try to parse bracketed data types like
this.  I deleted this customization SingleIdentifierGrammar in order to
avoid a very strange and incorrect change in the parse tree on one of
the test cases.  Otherwise there were no failures... so I think best
to delete this, and revisit any issue that comes up (this time, with
test cases demonstrating the problem).