Use the new CollationReferenceSegment everywhere
This is the result of searching for sequences of "COLLATE" followed by
some other segment type.  During this exercise, I realized that some
dialects also accept string literals instead.

This commit makes the CollationReferenceSegment in the ANSI dialect a
little extra permissive and accepts either an object reference or
a string literal.