Fix how keywords are defined
Follow example in other dialects to set the reserved and
non-reserved words, in a way that doesn't break code
coverage.

However, this resulted in a bunch of default things in the
parent ANSI dialect breaking, since Trino doesn't support
a lot of regular SQL (since it's mostly used for querying).