Improve support for Jinja templaters with custom tags
This would resolve #4624 by providing an easy way for plugins to provide
templaters that have custom Jinja tags/extensions loaded in their
environments.  The biggest problem is that JinjaAnalyzer has a lot of
hard-coded tag names in it.  This commit would allow a plugin author
to more easily support custom tags:

1.  The JinjaAnalyzer class now has a _get_tag_configuration function
    that returns a tag configuration for a given tag name.  Derived
    classes can override this function to provide additional behaviors
    for tag names.  The initial configuration provides a mapping of
    tag name to block type (e.g. block_start, block_mid, block_end).
    Plugins can also opt into the block tracing functionality.  For
    example, this can help ensure that a final tag at the end of the
    file without a trailing newline is preserved.

2.  The JinjaTemplater class did not provide an easy way for users to
    provide a derived JinjaAnalyzer class.  This class is refactored to
    add a new _get_jinja_analyzer function; this allows users to return
    a custom analyzer if they want.  This is similar in idea to the
    _get_jinja_env function, which allows returning a custom Jinja
    environment.

Some tests have also been included that test the behavior of a derived
JinjaAnalyzer and JinjaTemplater.  This also effectively provides a
small example of how to use it.

The provided test case also provides an example of why this commit is
useful.  Without it, the sample input string will contain two
"block_start" slices, and a "block_end": this is mismatched and causes
all kinds of problems.  With this commit, it's easy to configure the
analyzer so that the input is parsed with a single "block_start",
"block_mid", and "block_end".  That makes things work much better!