Upgraded dbt-core from 1.0 to 1.6 because of security
dbt-core below version 1.6 contains a library called 'werkzeug' which contains a 'high' risk vulnerability mentioned here: https://www.mend.io/vulnerability-database/CVE-2023-46136 
In dbt-core 1.6.0 this library was removed, as it was not used anyways, and thereby also removing the security thread. 
In big organizations, having these transient vulnerabilities in a repo is cause for for not allowing the use of the packages. Tools like whitesource/mend are oftenused to scan all repos to find and flag them for threads.