Clarify docs around subdir handling when loading macros
The documentation erroneously states that SQLFluff will not look in
subfolders when loading macros using the load_macros_from_path config.
This is not actually the case: the JinjaTemplater._extract_macros_from_path
function will recursively search subdirectories, and load those macros
into the Jinja global namespace as well.

This commit does the following:

1.  Amends one of the existing test cases to demonstrate that macros can
    be loaded from subdirectories.

2.  Updates the documentation to state that subdirectories will be
    searched as well.

There is no actual change in behavior of the product.  To the end user,
this is a doc correction.