Add loader_search_path setting to Jinja templater
Currently, in order to include a folder in the Jinja loader search path,
one must specify it using the load_macros_from_path setting.  However,
this setting carries a side effect, which may be unwanted in some
scenarios: it will also load every single file in that directory (which
is slow), and loads those macros into the global namespace (which may
not be an accurate representation of the end-user application behavior).
There is currently no option to avoid this additional behavior.

This commit adds a new loader_search_path setting, which is a simple way
of specifying the Jinja loader's search path.  The behavior of the
existing load_macros_from_path setting remains unchanged.  If both
options are set, all directories are merged and included in the Jinja
loader search path.