Add support for custom JinjaTracer implementations
Currently, the JinjaAnalyzer implementation will directly construct an
instance of JinjaTracer, making it difficult for users providing a
derived JinjaAnalyzer to also have it return a derived JinjaTracer.

By moving the construction of JinjaTracer to a separate function, it is
easy for plugin authors to override this function and provide their own
tracer.

This approach is very similar to the one taken for allowing custom
JinjaAnalyzer implementations in pull request #5543.